remove service test for responding to expired invite
There's no clean and simple way to update the expiration time of an
already-created Invite, or to set the exact expiration time in advance.
There's also no way to modify the behavior of `time.Now()` temporarily,
because the call to it happens inside the `auth.Service` method
(`.InviteRespond`) in this case.

Signed-off-by: Bogdan Mitrovic <bogdan.mitrovic@mainflux.com>